Move AWS Batch validators and tests in separate modules
