+++ cli/src/pcluster/validators/cluster_validators.py
@@ -12,11 +12,7 @@ import boto3
 from botocore.exceptions import ClientError
 
 from pcluster.models.common import DynamicParam, FailureLevel, Param, Validator
-from pcluster.utils import (
-    get_supported_architectures_for_instance_type,
-    get_supported_os_for_architecture,
-    is_instance_type_format,
-)
+from pcluster.utils import get_supported_architectures_for_instance_type, get_supported_os_for_architecture
 
 EFA_UNSUPPORTED_ARCHITECTURES_OSES = {
     "x86_64": [],
@@ -192,35 +188,3 @@ class InstanceArchitectureCompatibilityValidator(Validator):
                 ),
                 FailureLevel.ERROR,
             )
-
-
-class AwsbatchInstancesArchitectureCompatibilityValidator(Validator):
-    """Validate instance type and architecture combination."""
-
-    def _validate(self, instance_types: Param, architecture: DynamicParam):
-        """
-        Verify that head node and compute instance types imply compatible architectures.
-
-        When awsbatch is used as the scheduler, compute_instance_type can contain a CSV list.
-        """
-        head_node_architecture = architecture.value
-        for instance_type in instance_types.value.split(","):
-            # When awsbatch is used as the scheduler instance families can be used.
-            # Don't attempt to validate architectures for instance families, as it would require
-            # guessing a valid instance type from within the family.
-            if not is_instance_type_format(instance_type) and instance_type != "optimal":
-                self._add_failure(
-                    "Not validating architecture compatibility for compute instance type {0} because it does not have "
-                    "the expected format".format(instance_type),
-                    FailureLevel.INFO,
-                )
-                continue
-            compute_architectures = get_supported_architectures_for_instance_type(instance_type)
-            if head_node_architecture not in compute_architectures:
-                self._add_failure(
-                    "The specified compute instance type ({0}) supports the architectures {1}, none of which are "
-                    "compatible with the architecture supported by the head node instance type ({2}).".format(
-                        instance_type, compute_architectures, head_node_architecture
-                    ),
-                    FailureLevel.ERROR,
-                )
